requirements.txt
streamlit>=1.28.0
openai>=1.0.0
openai-agents>=0.0.14
pg8000>=1.29.0
python-dotenv>=0.19.0
yt-dlp
imagehash
Pillow
chromadb>=0.4.15
pathlib2 